Change import to follow file rename
