+++ qt/python/mantidqt/dialogs/test/test_spectraselectiondialog.py
@@ -24,7 +24,7 @@ from mantid.simpleapi import CreateSampleWorkspace, CropWorkspace
 from qtpy.QtWidgets import QDialogButtonBox
 
 # local imports
-from mantidqt.utils.qt.testing import requires_qapp
+from mantidqt.utils.qt.test import requires_qapp
 from mantidqt.dialogs.spectraselectordialog import parse_selection_str, SpectraSelectionDialog
