+++ qt/python/mantidqt/widgets/test/test_messagedisplay.py
@@ -20,7 +20,7 @@ from __future__ import (absolute_import, division, print_function,
 import unittest
 
 from mantidqt.widgets.messagedisplay import MessageDisplay
-from mantidqt.utils.qt.testing import requires_qapp
+from mantidqt.utils.qt.test import requires_qapp
 
 
 @requires_qapp
